pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "db-tools"
version = "0.1.0"
authors = [
  { name="Hoang Hong Huy", email="huy.hoanghong.work@gmail.com" },
]
description = "A CLI tool to seed and anonymize development databases."
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]",
    "rich",
    "SQLAlchemy",
    "Faker",
    "PyYAML",
    "textual"
]

[project.scripts]
db-tools = "db_tools.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest"]
postgres = ["psycopg2-binary"]
sqlserver = ["pyodbc"]